EPI-673: Remove reminder modal design updated
